backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djongo==1.3.6
pymongo==4.6.0
django-cors-headers==4.3.1
openai==1.3.7
python-dotenv==1.0.0
bson==0.5.10